requirements.txt
# AI Framework
pydantic-ai>=0.0.14
pydantic>=2.0.0

# Telegram Bot
python-telegram-bot[job-queue]>=22.5

# Database
psycopg[binary]>=3.1.0
psycopg-pool>=3.1.0
mem0ai>=0.1.0  # Semantic memory with pgvector backend

# Vision AI (choose one or both)
openai>=1.0.0
anthropic>=0.40.0

# Utilities
python-dotenv>=1.0.0
pillow>=10.0.0  # Image processing

# Development
pytest>=8.0.0
pytest-asyncio>=0.24.0
mypy>=1.13.0
ruff>=0.8.0
black>=24.0.0
